requirements.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
packaging==24.0
psycopg2==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
zipp==3.18.1
